+++ Firmware/USART.h
@@ -8,9 +8,9 @@ class USART
 		~USART();
 		static uint32_t getBaudRate();
 		bool hasRxData() const;
-		void read(uint8_t* out, uint8_t length) const;
+		void read(uint8_t* out, uint16_t length) const;
 		uint8_t readbyte() const;
-		void write(const uint8_t* in, uint8_t length) const;
+		void write(const uint8_t* in, uint16_t length) const;
 		void writebyte(uint8_t data) const;
 	private:
 		static uint32_t baudrate_;
